Split getComponent unit tests out
Over 500 lines for a unit test file is a bit unwieldy so it's now split into a file for each type of component.
