+++ src/utils/Component/getComponentTests/getComponent.phoneNumber.test.js
@@ -1,0 +1,52 @@
+// Global imports
+import { fireEvent, getAllByTestId, render } from '@testing-library/react';
+
+// Local imports
+import { ComponentTypes } from '../../../models';
+import getComponent from '../getComponent';
+
+describe('utils.Component.get', () => {
+
+  it('should return an appropriately rendered phone-number component', () => {
+    const ID = 'test-id';
+    const FIELD_ID = 'field-id';
+    const LABEL = 'label';
+    const ON_CHANGE_CALLS = [];
+    const ON_CHANGE = (e) => {
+      ON_CHANGE_CALLS.push(e.target);
+    };
+    const COMPONENT = {
+      type: ComponentTypes.PHONE_NUMBER,
+      id: ID,
+      fieldId: FIELD_ID,
+      label: LABEL,
+      onChange: ON_CHANGE,
+      'data-testid': ID
+    };
+    const { container } = render(getComponent(COMPONENT));
+
+    const [ formGroup, input ] = getAllByTestId(container, ID);
+    expect(formGroup.tagName).toEqual('DIV');
+    expect(formGroup.classList).toContain('govuk-form-group');
+    let label = undefined;
+    formGroup.childNodes.forEach(node => {
+      // Check if it's an element.
+      if (node instanceof Element && node.tagName === 'LABEL') {
+        label = node;
+      }
+    });
+    expect(label).toBeDefined();
+    expect(label.innerHTML).toContain(LABEL);
+    expect(label.getAttribute('for')).toEqual(ID);
+    expect(input.tagName).toEqual('INPUT');
+    expect(input.classList).toContain('govuk-input');
+    expect(input.id).toEqual(ID);
+    fireEvent.change(input, { target: { name: FIELD_ID, value: '07777' }});
+    expect(ON_CHANGE_CALLS.length).toEqual(1);
+    expect(ON_CHANGE_CALLS[0]).toMatchObject({
+      name: FIELD_ID,
+      value: '07777'
+    });
+  });
+
+});
